+++ coaster/utils/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+# flake8: noqa
 
 """
 Utilities
@@ -8,9 +9,11 @@ These functions are not dependent on Flask. They implement common patterns
 in Flask-based applications.
 """
 
+
 from __future__ import absolute_import
-from .misc import *  # NOQA
-from .text import *  # NOQA
-from .tsquery import *  # NOQA
-from .classes import *  # NOQA
-from ..shortuuid import suuid, encode as uuid2suuid, decode as suuid2uuid  # NOQA
+from .misc import *
+from .text import *
+from .markdown import *
+from .tsquery import *
+from .classes import *
+from ..shortuuid import suuid, encode as uuid2suuid, decode as suuid2uuid